EduMaterialModels

Documentation for EduMaterialModels.

EduMaterialModels.AbstractPlastic

EduMaterialModels.J2Plasticity

EduMaterialModels.J2ViscoPlasticity

EduMaterialModels.LinearIsotropicElasticity

EduMaterialModels.Zener1D

EduMaterialModels.simulate_response

EduMaterialModels.tosmandel

EduMaterialModels.tosvoigt

EduMaterialModels.AbstractPlastic — Type

AbstractPlastic

An AbstractPlastic materials have the same state variables, PlasticState, unknowns for the local equation system [Δλ, ϵp, κ, β], and von mises yield criterion. They may have different evolution laws.

EduMaterialModels.J2Plasticity — Type

J2Plasticity(;e, Y0, Hiso, κ∞, Hkin, β∞)

The J2Plasticity material model assumes linear elasticity combined with rate independent von mises plasticity. Nonlinear isotropic (Voce) and kinematic (Armstrong-Frederick) hardening are included as well.

EduMaterialModels.J2ViscoPlasticity — Type

J2ViscoPlasticity(;e, Y0, Hiso, κ∞, Hkin, β∞)

The J2ViscoPlasticity material model assumes linear elasticity combined with rate dependent (Norton-type) von mises plasticity. Nonlinear isotropic (Voce) and kinematic (Armstrong-Frederick) hardening are included as well. Specifically, the overstress function is

\[\eta(\Phi) = \left[ \frac{\langle \Phi \rangle}{Y_0}\right]^n\]

EduMaterialModels.LinearIsotropicElasticity — Type

LinearIsotropicElasticity(;kwargs...)

Construct a linear, isotropic, elastic material. The following pairs of properties can be given as keyword arguments:

G, K: Shear and bulk modulus

E, G: Young's and shear modulus

E, ν: Young's modulus and Poisson's ratio

E, K: Young's and bulk modulus

EduMaterialModels.Zener1D — Type

Zener1D(;E1, E2, η)

Special implementation of the Zener material model, which only works for uniaxial stress, i.e. when stress_state=MaterialModelsBase.UniaxialStress() is given to material_response.

EduMaterialModels.simulate_response — Method

A simple driver to simulate a response for a given material, m, a stress_state<:MaterialModelsBase.AbstractStressState, a strain history strain_history::Vector{<:AbstractTensor}, and time history, timevector::Vector{<:Number}.

EduMaterialModels.tosmandel — Method

tosmandel(A::Union{SecondOrderTensor,FourthOrderTensor})

SVector version of tomandel: converts the tensor A to a mandel (offdiagscale=√2 for symmetric tensors) SVector with the default voigt ordering. See also tosvoigt. The output can be converted back to a tensor with the standard frommandel function.

EduMaterialModels.tosvoigt — Method

tosvoigt(A::Union{SecondOrderTensor,FourthOrderTensor}; offdiagscale=1)

SVector version of tovoigt: converts the tensor A to a voigt SVector with the default voigt ordering. The keyword offdiagscale is only available for symmetric tensors. See also tosmandel. The output can be converted back to a tensor with the standard fromvoigt function.