Material models

EduMaterialModels.LinearIsotropicElasticity — Type

LinearIsotropicElasticity(;kwargs...)

Construct a linear, isotropic, elastic material. The following pairs of properties can be given as keyword arguments:

G, K: Shear and bulk modulus

E, G: Young's and shear modulus

E, ν: Young's modulus and Poisson's ratio

E, K: Young's and bulk modulus

EduMaterialModels.J2Plasticity — Type

J2Plasticity(;e, Y0, Hiso, κ∞, Hkin, β∞)

The J2Plasticity material model assumes linear elasticity combined with rate independent von mises plasticity. Nonlinear isotropic (Voce) and kinematic (Armstrong-Frederick) hardening are included as well.

EduMaterialModels.J2ViscoPlasticity — Type

J2ViscoPlasticity(;e, Y0, Hiso, κ∞, Hkin, β∞, n, tstar)

The J2ViscoPlasticity material model assumes linear elasticity combined with rate dependent (Norton-type) von mises plasticity. Nonlinear isotropic (Voce) and kinematic (Armstrong-Frederick) hardening are included as well. Specifically, the overstress function is

\[\eta(\Phi) = \left[ \frac{\langle \Phi \rangle}{Y_0}\right]^n\]

EduMaterialModels.Zener1D — Type

Zener1D(;E1, E2, η)

Special implementation of the Zener material model, which only works for uniaxial stress, i.e. when stress_state=MaterialModelsBase.UniaxialStress() is given to material_response.